gcov_map_srcfilename: do not read outside of new_suffix
Fix a mempcpy() call that was reading 6 bytes in the new_suffix string.
This string can be smaller so use strlen() and terminate with a '\0'.

This fixes the following GCC warning:

In function ‘gcov_map_srcfilename’,
    inlined from ‘gcov_match_line’ at cov-gcov.c:464:10:
cov-gcov.c:182:6: warning: ‘mempcpy’ forming offset [2, 6] is out of the bounds [0, 1] [-Warray-bounds]
  182 |  n = mempcpy(n, new_suffix, 6);
      |      ^~~~~~~~~~~~~~~~~~~~~~~~~

where new_suffix was the empty string.

Signed-off-by: Luc Michel <luc.michel@greensocs.com>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>